requirements.txt
openai~=0.28.0
python-dotenv~=1.0.1
sklearn
deep_translator
scikit-learn~=1.6.1
requests~=2.32.3
pygtfs
googlemaps~=4.10.0
torch~=2.6.0
sentence_transformers
fake_useragent
sentence-transformers~=3.4.1
transformers~=4.49.0
beautifulsoup4~=4.13.3
fake-useragent~=2.2.0
praw
matplotlib
streamlit
spacy
datasketch
